+++ app/assets/javascripts/blazer/application.js
@@ -9,10 +9,18 @@
 //= require ./moment
 //= require ./moment-timezone
 //= require ./daterangepicker
-//= require chartkick
+//= require ./chartkick
 //= require ./ace/ace
 //= require ./ace/ext-language_tools
 //= require ./ace/theme-twilight
 //= require ./ace/mode-sql
 //= require ./ace/snippets/text
 //= require ./ace/snippets/sql
+//= require ./Sortable
+//= require ./bootstrap
+
+$( function () {
+  $('.dropdown-toggle').mouseenter( function () {
+    $(this).parent().addClass('open');
+  });
+});