+++ app/models/blazer/dashboard_query.rb
@@ -1,0 +1,9 @@
+module Blazer
+  class DashboardQuery < ActiveRecord::Base
+    belongs_to :dashboard
+    belongs_to :query
+
+    validates :dashboard_id, presence: true
+    validates :query_id, presence: true
+  end
+end