+++ CHANGELOG.md
@@ -1,5 +1,43 @@
-## 0.0.6 [unreleased]
+## 1.0.3
 
+- Added maps
+- Added support for Rails 4.0
+
+## 1.0.2
+
+- Fixed error when installing
+- Added `schemas` option
+
+## 1.0.1
+
+- Added comments to queries
+- Added `cache` option
+- Added `user_method` option
+- Added `use_transaction` option
+
+## 1.0.0
+
+- Added support for multiple data sources
+- Added dashboards
+- Added checks
+- Added support for Redshift
+
+## 0.0.8
+
+- Easier to edit queries with variables
+- Dynamically expand editor height as needed
+- No need for spaces in search
+
+## 0.0.7
+
+- Fixed error when no `User` class
+- Fixed forking a query with variables
+- Set time zone after Rails initializes
+
+## 0.0.6
+
+- Added fork button
+- Fixed trending
 - Fixed time zones for date select
 
 ## 0.0.5